Fix download resumption by short-circuiting curl.
I assumed that `curl -C` would handle a complete file by reporting
a successful no-op, but that does not seem to be the case.

This fixes the problem by throwing away and redownloading the
signature, and not calling curl if the hash already checks out.

Can't add a test because it doesn't reproduce with file:// URLs.